fix(mcp): support OpenShell descriptor fallback
