+++ test/fixtures/langchain-deepagents-code/mcp_tools.py
@@ -1,0 +1,46 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Minimal pinned MCP loader fixture for the managed package patch tests."""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+
+def load_mcp_config(config_path):
+    path = Path(config_path)
+
+    if not path.exists():
+        error_msg = f"MCP config file not found: {config_path}"
+        raise FileNotFoundError(error_msg)
+
+    try:
+        with path.open(encoding="utf-8") as file_obj:
+            config = json.load(file_obj)
+    except json.JSONDecodeError:
+        raise
+    if "mcpServers" not in config:
+        raise ValueError("missing mcpServers")
+    return config
+
+
+async def resolve_and_load_mcp_tools(
+    *,
+    explicit_config_path=None,
+    project_context=None,
+):
+    configs = []
+    if explicit_config_path:
+        config_path = (
+            str(project_context.resolve_user_path(explicit_config_path))
+            if project_context is not None
+            else explicit_config_path
+        )
+        configs.append(load_mcp_config(config_path))
+    return configs
+
+
+def discover_mcp_configs(*, project_context=None):
+    del project_context
+    return [Path.home() / ".deepagents" / ".mcp.json"]
