+++ test/langchain-deepagents-code-direct-module-patch.test.ts
@@ -110,109 +110,10 @@ def cli_main():
   writeFixtureFile(
     packageDir,
     "app.py",
-    `
-from __future__ import annotations
-
-
-class UserMessage:
-    def __init__(self, value):
-        self.value = value
-
-
-class AppMessage(UserMessage):
-    pass
-
-
-class _Event:
-    def __init__(self):
-        self.was_set = False
-
-    def set(self):
-        self.was_set = True
-
-
-class DeepAgentsApp:
-    def __init__(self):
-        self.messages = []
-        self.notifications = []
-        self.original_commands = []
-        self.original_auth_manager = False
-        self.original_mcp_login = False
-        self.original_service_key = False
-        self.original_tavily = False
-        self.original_update_action = False
-        self.original_switch_kwargs = "not-called"
-        self._update_check_done = _Event()
-        self._auto_approve = True
-        self._status_bar = None
-        self._session_state = None
-        self._rubric_model = "attacker:model"
-        self._server_kwargs = {"rubric_model": "attacker:model"}
-
-    async def _mount_message(self, message):
-        self.messages.append(message.value)
-
-    def notify(self, message, **kwargs):
-        self.notifications.append((message, kwargs))
-
-    async def _handle_command(self, command):
-        self.original_commands.append(command)
-
-    async def _switch_model(self, model_spec, **kwargs):
-        del model_spec
-        self.original_switch_kwargs = kwargs.get("extra_kwargs")
-
-    async def _check_for_updates(self, *, periodic=False):
-        del periodic
-
-    async def _handle_update_command(self, command="/update"):
-        del command
-
-    async def _handle_install_command(self, command):
-        del command
-
-    async def _install_extra(self, *args, **kwargs):
-        del args, kwargs
-        return True
-
-    async def _handle_install_package(self, *args, **kwargs):
-        del args, kwargs
-
-    async def _handle_auto_update_toggle(self):
-        return None
-
-    async def _prompt_launch_tavily(self):
-        self.original_tavily = True
-
-    async def _prompt_model_auth_if_needed(self, model_spec):
-        del model_spec
-        return True
-
-    async def _show_auth_manager(self, **kwargs):
-        del kwargs
-        self.original_auth_manager = True
-
-    async def _enter_service_api_key(self, *args, **kwargs):
-        del args, kwargs
-        self.original_service_key = True
-
-    async def _handle_update_action(self, *args, **kwargs):
-        del args, kwargs
-        self.original_update_action = True
-
-    def _start_mcp_login(self, server_name):
-        del server_name
-        self.original_mcp_login = True
-
-    async def _on_auto_approve_enabled(self):
-        self._auto_approve = True
-
-    async def action_toggle_auto_approve(self):
-        self._auto_approve = not self._auto_approve
-
-    async def _set_rubric_model(self, model_spec):
-        self._rubric_model = model_spec
-`,
+    fs.readFileSync(
+      path.join(process.cwd(), "test", "fixtures", "langchain-deepagents-code", "app.py"),
+      "utf8",
+    ),
   );
   writeFixtureFile(
     packageDir,
@@ -339,48 +240,10 @@ def list_subagents(*args, **kwargs):
   writeFixtureFile(
     packageDir,
     "server.py",
-    `
-from __future__ import annotations
-
-import os
-import subprocess
-
-
-def _build_server_env():
-    return dict(os.environ)
-
-
-class ServerProcess:
-    def __init__(self, cmd, work_dir, env):
-        self.cmd = cmd
-        self.work_dir = work_dir
-        self.env = env
-        self.outputs = []
-        self._process = None
-
-    async def start(self):
-        cmd = self.cmd
-        work_dir = self.work_dir
-        env = self.env
-        self._log_file = subprocess.PIPE
-        self._process = subprocess.Popen(  # noqa: S603, ASYNC220
-            cmd,
-            cwd=str(work_dir),
-            env=env,
-            stdout=self._log_file,
-            stderr=subprocess.STDOUT,
-        )
-        output, _ = self._process.communicate(timeout=10)
-        if self._process.returncode != 0:
-            raise RuntimeError(output.decode())
-        self.outputs.append(output.decode())
-
-    async def restart(self):
-        if self._process is not None and self._process.poll() is None:
-            self._process.terminate()
-            self._process.wait(timeout=10)
-        await self.start()
-`,
+    fs.readFileSync(
+      path.join(process.cwd(), "test", "fixtures", "langchain-deepagents-code", "server.py"),
+      "utf8",
+    ),
   );
   writeFixtureFile(
     packageDir,
@@ -402,16 +265,10 @@ def _normalize_path(raw_path, project_context, label):
   writeFixtureFile(
     packageDir,
     "mcp_tools.py",
-    `
-from __future__ import annotations
-
-from pathlib import Path
-
-
-def discover_mcp_configs(*, project_context=None):
-    del project_context
-    return [Path.home() / ".deepagents" / ".mcp.json"]
-`,
+    fs.readFileSync(
+      path.join(process.cwd(), "test", "fixtures", "langchain-deepagents-code", "mcp_tools.py"),
+      "utf8",
+    ),
   );
   writeFixtureFile(
     packageDir,
@@ -904,8 +761,8 @@ describe("LangChain Deep Agents Code managed package patch", () => {
             "from pathlib import Path",
             "from deepagents_code import _nemoclaw_managed as managed",
             "managed._MCP_CONFIG_FILE = Path(sys.argv[1])",
-            "sealed = managed.managed_mcp_config_path()",
-            "print(Path(sealed).read_text(encoding='utf-8') if sealed else 'absent', end='')",
+            "snapshot = managed.managed_mcp_config_path()",
+            "print(managed.managed_mcp_config_bytes(snapshot).decode() if snapshot else 'absent', end='')",
           ].join("; "),
           configPath,
         ],
@@ -1043,7 +900,7 @@ describe("LangChain Deep Agents Code managed package patch", () => {
     expect(symlinked.status).not.toBe(0);
   });
 
-  it("passes one sealed managed MCP snapshot through ServerProcess.start and restart", () => {
+  it("passes sealed and anonymous MCP snapshots through ServerProcess restart", () => {
     const tempDir = createPackageFixture();
     patchFixture(tempDir);
     const configPath = path.join(tempDir, ".nemoclaw-mcp.json");
@@ -1058,14 +915,16 @@ describe("LangChain Deep Agents Code managed package patch", () => {
         },
       },
     };
-    fs.writeFileSync(configPath, `${JSON.stringify(managedConfig)}\n`, { mode: 0o600 });
+    for (const snapshotKind of ["sealed-memfd", "anonymous-otmpfile"] as const) {
+      fs.writeFileSync(configPath, `${JSON.stringify(managedConfig)}\n`, { mode: 0o600 });
 
-    const result = spawnSync(
-      "python3",
-      [
-        "-c",
-        `
+      const result = spawnSync(
+        "python3",
+        [
+          "-c",
+          `
 import asyncio
+import errno
 import fcntl
 import json
 import os
@@ -1073,39 +932,62 @@ import sys
 from pathlib import Path
 
 from deepagents_code import _nemoclaw_managed as managed
-from deepagents_code import _server_config, mcp_tools
+from deepagents_code import _server_config, app, mcp_tools
 from deepagents_code.server import ServerProcess
 
+real_memfd_create = os.memfd_create
+if sys.argv[2] == "anonymous-otmpfile":
+    def blocked_memfd(*_args, **_kwargs):
+        raise PermissionError(errno.EPERM, "blocked by seccomp")
+    managed.os.memfd_create = blocked_memfd
 managed._MCP_CONFIG_FILE = Path(sys.argv[1])
-sealed_path = managed.managed_mcp_config_path()
-assert sealed_path is not None
-descriptor = int(sealed_path.removeprefix("/proc/self/fd/"))
+snapshot_path = managed.managed_mcp_config_path()
+assert snapshot_path is not None
+descriptor = int(snapshot_path.removeprefix("/proc/self/fd/"))
+binding = managed._MANAGED_MCP_BINDING
+assert binding is not None
 required_seals = (
     fcntl.F_SEAL_WRITE
     | fcntl.F_SEAL_GROW
     | fcntl.F_SEAL_SHRINK
     | fcntl.F_SEAL_SEAL
 )
-assert fcntl.fcntl(descriptor, fcntl.F_GET_SEALS) == required_seals
-assert _server_config._normalize_path(sealed_path, None, "MCP config") == sealed_path
+if binding["kind"] == managed._MCP_SEALED_KIND:
+    assert fcntl.fcntl(descriptor, fcntl.F_GET_SEALS) == required_seals
+else:
+    assert binding["kind"] == managed._MCP_ANONYMOUS_KIND
+    assert fcntl.fcntl(descriptor, fcntl.F_GETFL) & os.O_ACCMODE == os.O_RDONLY
+assert managed.managed_mcp_config_bytes(snapshot_path) == managed.managed_mcp_config_bytes(snapshot_path)
+assert _server_config._normalize_path(snapshot_path, None, "MCP config") == snapshot_path
+assert app.DeepAgentsApp._absolutize_launch_relative_path(
+    snapshot_path, Path.cwd()
+) == snapshot_path
 assert mcp_tools.discover_mcp_configs() == []
+expected_config = json.loads(managed.managed_mcp_config_bytes(snapshot_path))
+
+class RejectingProjectContext:
+    def resolve_user_path(self, _path):
+        raise AssertionError("managed descriptor path must not be resolved")
 
 child = (
-    "import os; from pathlib import Path; "
-    "print(Path(os.environ['DEEPAGENTS_CODE_SERVER_MCP_CONFIG_PATH']).read_text(), end='')"
+    "import json, os; from deepagents_code.mcp_tools import load_mcp_config; "
+    "config = load_mcp_config(os.environ['DEEPAGENTS_CODE_SERVER_MCP_CONFIG_PATH']); "
+    "assert 'NEMOCLAW_DCODE_MCP_BINDING' not in os.environ; "
+    "print(json.dumps(config), end='')"
 )
 def server_for_path(config_path):
     env = os.environ.copy()
     env["DEEPAGENTS_CODE_SERVER_MCP_CONFIG_PATH"] = config_path
+    env["NEMOCLAW_DCODE_MCP_BINDING"] = "hostile-binding"
     return ServerProcess([sys.executable, "-c", child], os.getcwd(), env)
 
 def make_descriptor_server(name, payload, seals):
-    descriptor = os.memfd_create(name, flags=os.MFD_ALLOW_SEALING)
+    descriptor = real_memfd_create(name, flags=os.MFD_ALLOW_SEALING)
     os.write(descriptor, payload)
     fcntl.fcntl(descriptor, fcntl.F_ADD_SEALS, seals)
     return descriptor, server_for_path(f"/proc/self/fd/{descriptor}")
 
-server = server_for_path(sealed_path)
+server = server_for_path(snapshot_path)
 unsealed_descriptor, unsealed_server = make_descriptor_server(
     "unsealed-dcode-mcp", b"{}", 0
 )
@@ -1117,6 +999,11 @@ oversized_descriptor, oversized_server = make_descriptor_server(
 )
 
 async def exercise():
+    resolved_configs = await mcp_tools.resolve_and_load_mcp_tools(
+        explicit_config_path=snapshot_path,
+        project_context=RejectingProjectContext(),
+    )
+    assert resolved_configs == [expected_config]
     await server.start()
     Path(sys.argv[1]).write_text(
         json.dumps({
@@ -1137,7 +1024,8 @@ async def exercise():
         try:
             await invalid_server.start()
         except RuntimeError as exc:
-            assert "not valid and sealed" in str(exc)
+            assert "not process-local" in str(exc)
+            assert not hasattr(invalid_server, "_log_file")
         else:
             raise AssertionError("invalid MCP descriptor was inherited")
 
@@ -1145,27 +1033,32 @@ asyncio.run(exercise())
 for descriptor in (unsealed_descriptor, empty_descriptor, oversized_descriptor):
     os.close(descriptor)
 print(json.dumps({
-    "path": sealed_path,
+    "path": snapshot_path,
+    "kind": binding["kind"],
     "outputs": [json.loads(output) for output in server.outputs],
 }))
 `,
-        configPath,
-      ],
-      {
-        cwd: tempDir,
-        env: { PATH: process.env.PATH, PYTHONPATH: tempDir },
-        encoding: "utf8",
-      },
-    );
+          configPath,
+          snapshotKind,
+        ],
+        {
+          cwd: tempDir,
+          env: { PATH: process.env.PATH, PYTHONPATH: tempDir },
+          encoding: "utf8",
+        },
+      );
 
-    expect(result.status, result.stderr).toBe(0);
-    const proof = JSON.parse(result.stdout) as {
-      path: string;
-      outputs: unknown[];
-    };
-    expect(proof.path).toMatch(/^\/proc\/self\/fd\/[0-9]+$/);
-    expect(proof.outputs).toEqual([managedConfig, managedConfig]);
-    expect(result.stdout).not.toContain("attacker");
+      expect(result.status, result.stderr).toBe(0);
+      const proof = JSON.parse(result.stdout) as {
+        path: string;
+        kind: string;
+        outputs: unknown[];
+      };
+      expect(proof.path).toMatch(/^\/proc\/self\/fd\/[0-9]+$/);
+      expect(proof.kind).toBe(snapshotKind);
+      expect(proof.outputs).toEqual([managedConfig, managedConfig]);
+      expect(result.stdout).not.toContain("attacker");
+    }
   });
 
   it("blocks TUI commands, credential screens, dotenv, OAuth, and install backends", () => {
@@ -1373,12 +1266,15 @@ async def validate():
     assert headless_kwargs["rubric_model"] is None
     assert non_interactive.settings.shell_allow_list is None
     _nemoclaw_managed._MCP_CONFIG_FILE = Path(${JSON.stringify(managedMcpPath)})
-    _nemoclaw_managed._MANAGED_MCP_FD = None
+    _nemoclaw_managed._MANAGED_MCP_FD = _nemoclaw_managed._MANAGED_MCP_BINDING = None
     _nemoclaw_managed._MANAGED_MCP_READY = False
     managed_args = dcode_main.parse_args()
-    sealed_mcp_path = managed_args.mcp_config
-    assert sealed_mcp_path.startswith("/proc/self/fd/")
-    assert Path(sealed_mcp_path).is_file()
+    snapshot_mcp_path = managed_args.mcp_config
+    assert snapshot_mcp_path.startswith("/proc/self/fd/")
+    assert Path(snapshot_mcp_path).is_file()
+    assert instance._absolutize_launch_relative_path(
+        snapshot_mcp_path, Path.cwd()
+    ) == snapshot_mcp_path
     assert managed_args.no_mcp is False
     assert managed_args.trust_project_mcp is False
     managed_headless_kwargs = await non_interactive.run_non_interactive(
@@ -1388,7 +1284,7 @@ async def validate():
         no_mcp=True,
         trust_project_mcp=True,
     )
-    assert managed_headless_kwargs["mcp_config_path"] == sealed_mcp_path
+    assert managed_headless_kwargs["mcp_config_path"] == snapshot_mcp_path
     assert managed_headless_kwargs["no_mcp"] is False
     assert managed_headless_kwargs["trust_project_mcp"] is False
     assert model_config.ModelConfig().get_class_path("openai") is None
